Current user checks & some refacto

+++ internal/api/workout_handler.go
@@ -27,20 +27,20 @@ func (wh *WorkoutHandler) GetWorkout(w http.ResponseWriter, r *http.Request) {
 	workoutID, err := utils.ParseIDParamFromURL(r, "workoutId")
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "invalid ID param"})
+		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "bad request"})
 		return
 	}
 
 	workout, err := wh.workoutStore.GetWorkout(workoutID)
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed to retrieve resource"})
+		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed"})
 		return
 	}
 
 	if workout == nil {
 		wh.logger.Printf("ERROR: workout %d not found", workoutID)
-		_ = utils.WriteJSONResponse(w, http.StatusNotFound, utils.Envelope{"error": "resource not found"})
+		_ = utils.WriteJSONResponse(w, http.StatusNotFound, utils.Envelope{"error": "not found"})
 		return
 	}
 
@@ -53,14 +53,17 @@ func (wh *WorkoutHandler) CreateWorkout(w http.ResponseWriter, r *http.Request) 
 	err := json.NewDecoder(r.Body).Decode(&workout)
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "invalid payload"})
+		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "bad request"})
 		return
 	}
+
+	currentUser := r.Context().Value("user").(store.User)
+	workout.UserID = currentUser.ID
 
 	err = wh.workoutStore.PersistWorkout(&workout)
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed to persist resource"})
+		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed"})
 		return
 	}
 
@@ -71,20 +74,20 @@ func (wh *WorkoutHandler) UpdateWorkout(w http.ResponseWriter, r *http.Request) 
 	workoutID, err := utils.ParseIDParamFromURL(r, "workoutId")
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "invalid ID param"})
+		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "bad request"})
 		return
 	}
 
 	existingWorkout, err := wh.workoutStore.GetWorkout(workoutID)
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed to retrieve resource"})
+		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed"})
 		return
 	}
 
 	if existingWorkout == nil {
 		wh.logger.Printf("ERROR: workout %d not found", workoutID)
-		_ = utils.WriteJSONResponse(w, http.StatusNotFound, utils.Envelope{"error": "resource not found"})
+		_ = utils.WriteJSONResponse(w, http.StatusNotFound, utils.Envelope{"error": "not found"})
 		return
 	}
 
@@ -99,7 +102,7 @@ func (wh *WorkoutHandler) UpdateWorkout(w http.ResponseWriter, r *http.Request) 
 	err = json.NewDecoder(r.Body).Decode(&updateWorkoutPayload)
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "invalid payload"})
+		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "bad request"})
 		return
 	}
 
@@ -119,10 +122,28 @@ func (wh *WorkoutHandler) UpdateWorkout(w http.ResponseWriter, r *http.Request) 
 		existingWorkout.Exercises = updateWorkoutPayload.Exercises
 	}
 
+	currentUser := r.Context().Value("user").(store.User)
+	workoutOwner, err := wh.workoutStore.GetWorkoutOwner(workoutID)
+	if errors.Is(err, sql.ErrNoRows) {
+		wh.logger.Printf("ERROR: %v", err)
+		_ = utils.WriteJSONResponse(w, http.StatusNotFound, utils.Envelope{"error": "not found"})
+		return
+	}
+	if err != nil {
+		wh.logger.Printf("ERROR: %v", err)
+		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed"})
+		return
+	}
+	if workoutOwner != currentUser.ID {
+		wh.logger.Printf("ERROR: %v", err)
+		_ = utils.WriteJSONResponse(w, http.StatusForbidden, utils.Envelope{"error": "forbidden"})
+		return
+	}
+
 	err = wh.workoutStore.UpdateWorkout(existingWorkout)
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed to update resource"})
+		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed"})
 		return
 	}
 
@@ -133,20 +154,38 @@ func (wh *WorkoutHandler) DeleteWorkout(w http.ResponseWriter, r *http.Request) 
 	workoutID, err := utils.ParseIDParamFromURL(r, "workoutId")
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "invalid ID param"})
+		_ = utils.WriteJSONResponse(w, http.StatusBadRequest, utils.Envelope{"error": "bad request"})
 		return
 	}
 
-	err = wh.workoutStore.DeleteWorkout(workoutID)
-	// idempotency?
+	currentUser := r.Context().Value("user").(store.User)
+	workoutOwner, err := wh.workoutStore.GetWorkoutOwner(workoutID)
 	if errors.Is(err, sql.ErrNoRows) {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusNotFound, utils.Envelope{"error": "resource not found"})
+		_ = utils.WriteJSONResponse(w, http.StatusNotFound, utils.Envelope{"error": "not found"})
 		return
 	}
 	if err != nil {
 		wh.logger.Printf("ERROR: %v", err)
-		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed to delete resource"})
+		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed"})
+		return
+	}
+	if workoutOwner != currentUser.ID {
+		wh.logger.Printf("ERROR: %v", err)
+		_ = utils.WriteJSONResponse(w, http.StatusForbidden, utils.Envelope{"error": "forbidden"})
+		return
+	}
+
+	err = wh.workoutStore.DeleteWorkout(workoutID)
+	// Question: Idempotency?
+	if errors.Is(err, sql.ErrNoRows) {
+		wh.logger.Printf("ERROR: %v", err)
+		_ = utils.WriteJSONResponse(w, http.StatusNotFound, utils.Envelope{"error": "not found"})
+		return
+	}
+	if err != nil {
+		wh.logger.Printf("ERROR: %v", err)
+		_ = utils.WriteJSONResponse(w, http.StatusInternalServerError, utils.Envelope{"error": "failed"})
 		return
 	}
